Expose Current bindable in LabelledComponents
Adds a `LabelledDrawable` class for usages where bindables are not present.
